+++ app/web-components/src/client/preview/types.ts
@@ -1,5 +1,4 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { TemplateResult, SVGTemplateResult } from 'lit-element';
+import { TemplateResult, SVGTemplateResult } from 'lit-html';
 
 export type { RenderContext } from '@storybook/core';
 export { Args, ArgTypes, Parameters, StoryContext } from '@storybook/addons';