+++ lib/builder-webpack4/src/preview/iframe-webpack.config.ts
@@ -124,13 +124,13 @@ export default async ({
     // stats: 'errors-only',
     output: {
       path: path.resolve(process.cwd(), outputDir),
-      filename: '[name].[hash].bundle.js',
+      filename: isProd ? '[name].[contenthash:8].iframe.bundle.js' : '[name].iframe.bundle.js',
       publicPath: '',
     },
-    // watchOptions: {
-    //   aggregateTimeout: 10,
-    //   ignored: /node_modules/,
-    // },
+    watchOptions: {
+      aggregateTimeout: 10,
+      ignored: /node_modules/,
+    },
     plugins: [
       new FilterWarningsPlugin({
         exclude: /export '\S+' was not found in 'global'/,
@@ -194,7 +194,7 @@ export default async ({
     resolve: {
       extensions: ['.mjs', '.js', '.jsx', '.ts', '.tsx', '.json', '.cjs'],
       modules: ['node_modules'].concat(envs.NODE_PATH || []),
-      mainFields: isProd ? undefined : ['browser', 'main'],
+      mainFields: ['browser', 'module', 'main'],
       alias: {
         ...themingPaths,
         ...storybookPaths,