+++ lib/builder-webpack5/src/preview/base-webpack.config.ts
@@ -1,5 +1,5 @@
 import { logger } from '@storybook/node-logger';
-import type { Options } from '@storybook/core-common';
+import type { Options, BuilderOptions } from '@storybook/core-common';
 import type { Configuration } from 'webpack';
 
 export async function createDefaultWebpackConfig(
@@ -27,7 +27,7 @@ export async function createDefaultWebpackConfig(
       test: /\.css$/,
       sideEffects: true,
       use: [
-        // TODO(blaine): Decide if we want to keep style-loader & css-loader in core
+        // TODO: Decide if we want to keep style-loader & css-loader in core
         // Trying to apply style-loader or css-loader to files that already have been
         // processed by them causes webpack to crash, so no one else can add similar
         // loader configurations to the `.css` extension.
@@ -42,6 +42,8 @@ export async function createDefaultWebpackConfig(
     };
   }
 
+  const isProd = storybookBaseConfig.mode !== 'development';
+
   return {
     ...storybookBaseConfig,
     module: {
@@ -53,7 +55,9 @@ export async function createDefaultWebpackConfig(
           test: /\.(svg|ico|jpg|jpeg|png|apng|gif|eot|otf|webp|ttf|woff|woff2|cur|ani|pdf)(\?.*)?$/,
           loader: require.resolve('file-loader'),
           options: {
-            name: 'static/media/[name].[hash:8].[ext]',
+            name: isProd
+              ? 'static/media/[name].[contenthash:8].[ext]'
+              : 'static/media/[path][name].[ext]',
           },
         },
         {
@@ -61,7 +65,9 @@ export async function createDefaultWebpackConfig(
           loader: require.resolve('url-loader'),
           options: {
             limit: 10000,
-            name: 'static/media/[name].[hash:8].[ext]',
+            name: isProd
+              ? 'static/media/[name].[contenthash:8].[ext]'
+              : 'static/media/[path][name].[ext]',
           },
         },
       ],