+++ lib/core-common/typings.d.ts
@@ -1,6 +1,6 @@
 declare module 'lazy-universal-dotenv';
 declare module 'pnp-webpack-plugin';
-
+declare module '@storybook/semver';
 declare module 'file-system-cache' {
   export interface Options {
     basePath?: string;
@@ -15,12 +15,12 @@ declare module 'file-system-cache' {
     ensureBasePath(): Promise<void>;
     get(key: string, defaultValue?: any): Promise<any | typeof defaultValue>;
     getSync(key: string, defaultValue?: any): any | typeof defaultValue;
-    set(key: string, value: any): Promise<{ path: string }>
+    set(key: string, value: any): Promise<{ path: string }>;
     setSync(key: string, value: any): this;
     remove(key: string): Promise<void>;
     clear(): Promise<void>;
     save(): Promise<{ paths: string[] }>;
-    load(): Promise<{ files: Array<{ path: string, value: any }> }>;
+    load(): Promise<{ files: Array<{ path: string; value: any }> }>;
   }
 
   function create(options: Options): FileSystemCache;